requirements.txt
Django==3.1.3
djangorestframework_simplejwt==4.5.0
python-decouple==3.3
djangorestframework==3.12.1
psycopg2==2.8.6